CodeJar experiment
CodeJar is a minimal web-based text editor, much lighter than Monaco
(the VSCode editor). Rather than defining a grammar for the syntax,
CodeJar allows us to use our own tokenizer and parser. Currently,
only the tokenizer is used, but soon it will be replaced by a more
fault-tolerant parser than we currently have.
