+++ apps/playground/monaco.html
@@ -2,6 +2,7 @@
 <html>
 <head>
     <title>uFork Playground</title>
+    <meta name="viewport" content="width=device-width" />
     <link
         rel="stylesheet"
         href="https://esm.sh/monaco-editor@0.45.0/dev/vs/editor/editor.main.css"
@@ -26,6 +27,6 @@
 </head>
 <body>
     <div id="editor"></div>
-    <script type="module" src="main.js"></script>
+    <script type="module" src="monaco.js"></script>
 </body>
 </html>
