+++ lib/assemble.js
@@ -6,204 +6,7 @@
 // The assembly language is described in asm.md.
 // The intermediate representation is described in ir.md.
 
-// Tokenizer ///////////////////////////////////////////////////////////////////
-
-function tag_regexp(strings) {
-
-// A tag function that creates a RegExp from a template literal string. Any
-// whitespace in the string is ignored, and so can be injected into the pattern
-// to improve readability.
-
-    return new RegExp(strings.raw[0].replace(/\s/g, ""), "");
-}
-
-const rx_token_raw = tag_regexp `
-    (
-        \n
-      | \r \n?
-    )
-  | ( \u0020+ )
-  | ( ; .* )
-  | (
-        [ a-z A-Z ]
-        (?:
-            [ \- _ ]? [ 0-9 a-z A-Z ]
-        )*
-    )
-  | (
-        #
-        [ a-z _ ? ]+
-    )
-  | (
-        0
-      | -? [ 1-9 ] \d* (?: # [ a-z A-Z \d ]+ )?
-    )
-  | (
-        "
-        [ ^ " ]*
-        "
-    )
-  | (
-        [ \. : # \? ]
-    )
-  | (
-        '
-        (?:
-            \\ [ \\ ' b t n r ]
-          | [ ^ \\ ' ]
-        )
-        '
-    )
-`;
-
-// Capturing groups:
-//  [1] Newline
-//  [2] Whitespace
-//  [3] Comment
-//  [4] Name
-//  [5] Literal
-//  [6] Fixnum
-//  [7] String
-//  [8] Punctuator
-//  [9] Character
-
-const escape_code_points = {
-    "\\": 0x5C,
-    "'": 0x27,
-    "b": 0x08,
-    "t": 0x09,
-    "n": 0x0A,
-    "r": 0x0D
-};
-
-function tokenize(source) {
-    let rx_token = new RegExp(rx_token_raw, "yu"); // sticky, unicode aware
-    let position = 0;
-    return function token_generator() {
-
-        function error() {
-            source = undefined;
-            return {
-                id: ":error:",
-                start: position
-            };
-        }
-
-        if (source === undefined) {
-            return error();
-        }
-        if (rx_token.lastIndex >= source.length) {
-            return;
-        }
-        let captives = rx_token.exec(source);
-        if (!captives) {
-            return error();
-        }
-
-// The following code is incorrect. The "start" and "end" positions should be
-// measured in Unicode code points, because the intermediate representation is
-// intended to be language independent. For now, for a simple implementation,
-// we count UTF-16 character codes instead.
-
-        const start = position;
-        const end = position + captives[0].length;
-        position = end;
-        if (captives[1]) {
-            return {
-                id: ":linebreak:",
-                start,
-                end
-            };
-        }
-        if (captives[2]) {
-            return {
-                id: ":space:",
-                start,
-                end
-            };
-        }
-        if (captives[3]) {
-            return {
-                id: ":comment:",
-                comment: captives[3].slice(1),
-                start,
-                end
-            };
-        }
-        if (captives[4]) {
-            return {
-                id: ":name:",
-                text: captives[4],
-                start,
-                end
-            };
-        }
-        if (captives[5]) {
-            return {
-                id: ":literal:",
-                name: captives[5].slice(1),
-                start,
-                end
-            };
-        }
-        if (captives[6]) {
-            let [base, digits] = captives[6].split("#");
-            if (digits === undefined) {
-                digits = base;
-                base = 10;
-            }
-            if (base < 0) {
-                return error();
-            }
-            const number = parseInt(digits, base);
-            return (
-                Number.isSafeInteger(number)
-                ? {
-                    id: ":number:",
-                    number,
-                    text: captives[6],
-                    start,
-                    end
-                }
-                : error()
-            );
-        }
-        if (captives[7]) {
-            return {
-                id: ":string:",
-                text: captives[7].slice(1, -1),
-                start,
-                end
-            };
-        }
-        if (captives[8]) {
-            return {
-                id: captives[8],
-                start,
-                end
-            };
-        }
-        if (captives[9]) {
-            const character = captives[9].slice(1, -1);
-            const code_point = (
-                character.startsWith("\\")
-                ? escape_code_points[character[1]]
-                : character.codePointAt(0)
-            );
-            return (
-                Number.isSafeInteger(code_point)
-                ? {
-                    id: ":number:",
-                    number: code_point,
-                    text: character,
-                    start,
-                    end
-                }
-                : error()
-            );
-        }
-    };
-}
+import tokenize from "./asm_tokenize.js";
 
 // Parser //////////////////////////////////////////////////////////////////////
